groupId changed to com.miqdigital.automation
