--- heimdall.application/src/main/java/com/miq/cucumber_runner/BeforeSuite.java
@@ -1,15 +1,0 @@
-package com.miq.cucumber_runner;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * The interface Before suite.
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
-public @interface BeforeSuite {
-
-}
