--- heimdall.application/src/main/java/com/miq/execution/dto/ExecutionInfo.java
@@ -1,30 +1,0 @@
-package com.miq.execution.dto;
-
-import java.util.Arrays;
-import java.util.List;
-
-import com.miq.scenario.dto.ScenarioInfo;
-
-import lombok.Builder;
-
-@Builder
-public class ExecutionInfo {
-  public String environment;
-  public String testType;
-  public String dateTime;
-  public long totalTests;
-  public long passTestCount;
-  public long failTestCount;
-  public String BuildName;
-  public String BuildNumber;
-  public List<ScenarioInfo> scenarioInfoList;
-
-  @Override
-  public String toString() {
-    return "ExecutionInfo{" + "environment='" + environment + '\'' + ", testType='" + testType
-        + '\'' + ", dateTime='" + dateTime + '\'' + ", totalTests=" + totalTests
-        + ", passTestCount=" + passTestCount + ", failTestCount=" + failTestCount + ", BuildName='"
-        + BuildName + '\'' + ", BuildNumber='" + BuildNumber + '\'' + ", scenarioInfoList=" + Arrays
-        .toString(scenarioInfoList.toArray()) + '}';
-  }
-}
