--- heimdall.application/src/main/java/com/miq/scenario/dto/ScenarioInfo.java
@@ -1,17 +1,0 @@
-package com.miq.scenario.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class ScenarioInfo {
-  private String scenarioTagId;
-  private String scenarioName;
-  private String scenarioStatus;
-  private String scenarioTotalDuration;
-  private int scenarioTotalSteps;
-  private String featureDescription;
-  private String featureFile;
-  private ScenarioStep scenarioStep;
-}
