--- heimdall.application/src/main/java/com/miq/scenario/dto/ScenarioStep.java
@@ -1,10 +1,0 @@
-package com.miq.scenario.dto;
-
-import lombok.Builder;
-
-@Builder
-public class ScenarioStep {
-  private String errMessage;
-  private String stepDuration;
-  private String scenarioLine;
-}
