--- heimdall.application/src/main/java/com/miq/slack/dto/SlackChannelInfo.java
@@ -1,10 +1,0 @@
-package com.miq.slack.dto;
-
-import lombok.Builder;
-
-@Builder
-public class SlackChannelInfo {
-  public String channelName;
-  public String token;
-  public boolean isNotifySlack;
-}
