--- heimdall.application/src/main/java/com/miq/slack/dto/SlackMessageInfo.java
@@ -1,10 +1,0 @@
-package com.miq.slack.dto;
-
-import lombok.Builder;
-
-@Builder
-public class SlackMessageInfo {
-  public StringBuilder testExecutionInfo;
-  public StringBuilder failedTestDescription;
-  public long failedTestCount;
-}
