+++ heimdall.application/src/main/java/com/miqdigital/scenario/dto/ScenarioInfo.java
@@ -1,0 +1,17 @@
+package com.miqdigital.scenario.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class ScenarioInfo {
+  private String scenarioTagId;
+  private String scenarioName;
+  private String scenarioStatus;
+  private String scenarioTotalDuration;
+  private int scenarioTotalSteps;
+  private String featureDescription;
+  private String featureFile;
+  private ScenarioStep scenarioStep;
+}
