+++ heimdall.application/src/main/java/com/miqdigital/scenario/dto/ScenarioStep.java
@@ -1,0 +1,10 @@
+package com.miqdigital.scenario.dto;
+
+import lombok.Builder;
+
+@Builder
+public class ScenarioStep {
+  private String errMessage;
+  private String stepDuration;
+  private String scenarioLine;
+}
