+++ heimdall.application/src/main/java/com/miqdigital/slack/dto/SlackChannelInfo.java
@@ -1,0 +1,10 @@
+package com.miqdigital.slack.dto;
+
+import lombok.Builder;
+
+@Builder
+public class SlackChannelInfo {
+  public String channelName;
+  public String token;
+  public boolean isNotifySlack;
+}
