+++ heimdall.application/src/main/java/com/miqdigital/slack/dto/SlackMessageInfo.java
@@ -1,0 +1,10 @@
+package com.miqdigital.slack.dto;
+
+import lombok.Builder;
+
+@Builder
+public class SlackMessageInfo {
+  public StringBuilder testExecutionInfo;
+  public StringBuilder failedTestDescription;
+  public long failedTestCount;
+}
