+++ heimdall.sample-client/pom.xml
@@ -5,7 +5,7 @@
 
     <parent>
         <artifactId>heimdall</artifactId>
-        <groupId>com.miq.automation</groupId>
+        <groupId>com.miqdigital.automation</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
@@ -33,7 +33,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.miq.automation</groupId>
+            <groupId>com.miqdigital.automation</groupId>
             <artifactId>heimdall.application</artifactId>
             <version>${heimdall.application.version}</version>
         </dependency>
